requirements.txt
asgiref==3.2.10
attrs==20.3.0
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-jwt-auth==0.2.0
google-auth==1.23.0
idna==2.10
importlib-metadata==3.1.0
iniconfig==1.1.1
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.1
packaging==20.7
Pillow==8.0.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.1
requests==2.25.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
zipp==3.4.0
